星旅完整功能

+++ src/main/java/com/drag/cstgroup/user/entity/User.java
@@ -121,7 +121,7 @@ public class User implements Serializable {
 	/**
 	 * 总充值金额
 	 */
-	private int rechargeBalance;
+	private BigDecimal rechargeBalance;
 	
 	//-----------企业信息------------
 	/**
