+++ src/main/java/com/drag/cstgroup/user/vo/UserVo.java
@@ -91,7 +91,7 @@ public class UserVo implements Serializable {
 	/**
 	 * 总充值金额
 	 */
-	private int rechargeBalance;
+	private BigDecimal rechargeBalance;
 	
 	//-----------企业信息------------
 	/**
